Macro gc::custom_trace

macro_rules! custom_trace {
    ($this:ident, $body:expr) => { ... };
}

This rule implements the trace method.

You define a this parameter name and pass in a body, which should call mark on every traceable element inside the body. The mark implementation will automatically delegate to the correct method on the argument.