Struct gc::Gc

pub struct Gc<T: Trace + ?Sized + 'static> { /* fields omitted */ }

A garbage-collected pointer type over an immutable value.

See the module level documentation for more details.

Implementations

impl<T: Trace> Gc<T>

pub fn new(value: T) -> Self

Constructs a new Gc<T> with the given value.

Collection

This method could trigger a garbage collection.

Examples

use gc::Gc;

let five = Gc::new(5);
assert_eq!(*five, 5);

impl<T: Trace + ?Sized> Gc<T>

pub fn ptr_eq(this: &Gc<T>, other: &Gc<T>) -> bool

Returns true if the two Gcs point to the same allocation.

impl<T: Trace + ?Sized> Gc<T>

pub fn into_raw(this: Self) -> *const T

Consumes the Gc, returning the wrapped pointer.

To avoid a memory leak, the pointer must be converted back into a Gc using Gc::from_raw.

Examples

use gc::Gc;

let x = Gc::new(22);
let x_ptr = Gc::into_raw(x);
assert_eq!(unsafe { *x_ptr }, 22);

pub unsafe fn from_raw(ptr: *const T) -> Self

Constructs an Gc from a raw pointer.

The raw pointer must have been previously returned by a call to a Gc::into_raw.

This function is unsafe because improper use may lead to memory problems. For example, a use-after-free will occur if the function is called twice on the same raw pointer.

Examples

use gc::Gc;

let x = Gc::new(22);
let x_ptr = Gc::into_raw(x);

unsafe {
    // Convert back to an `Gc` to prevent leak.
    let x = Gc::from_raw(x_ptr);
    assert_eq!(*x, 22);

    // Further calls to `Gc::from_raw(x_ptr)` would be memory unsafe.
}

// The memory can be freed at any time after `x` went out of scope above
// (when the collector is run), which would result in `x_ptr` dangling!