Crate gc

Thread-local garbage-collected boxes (The Gc<T> type).

The Gc<T> type provides shared ownership of an immutable value. It is marked as non-sendable because the garbage collection only occurs thread-locally.

Macros

custom_trace
This rule implements the trace method.

unsafe_empty_trace
This rule implements the trace methods with empty implementations.

Structs

Gc
A garbage-collected pointer type over an immutable value.

GcCell
A mutable memory location with dynamically checked borrow rules that can be used inside of a garbage-collected pointer.

GcCellRef
A wrapper type for an immutably borrowed value from a GcCell<T>.

GcCellRefMut
A wrapper type for a mutably borrowed value from a GcCell<T>.

Traits

Finalize
The Finalize trait. Can be specialized for a specific type to define finalization logic for that type.

Trace
The Trace trait, which needs to be implemented on garbage-collected objects.

Functions

finalizer_safe

force_collect
Immediately triggers a garbage collection on the current thread.