Struct gc::GcCell

pub struct GcCell<T: ?Sized + 'static> { /* fields omitted */ }

A mutable memory location with dynamically checked borrow rules that can be used inside of a garbage-collected pointer.

This object is a RefCell that can be used inside of a Gc<T>.

Methods

impl<T: Trace> GcCell<T>

pub fn new(value: T) -> Self

Creates a new GcCell containing value.

pub fn into_inner(self) -> T

Consumes the GcCell, returning the wrapped value.

impl<T: Trace + ?Sized> GcCell<T>

pub fn borrow(&self) -> GcCellRef<T>

Immutably borrows the wrapped value.

The borrow lasts until the returned GcCellRef exits scope. Multiple immutable borrows can be taken out at the same time.

Panics

Panics if the value is currently mutably borrowed.

pub fn borrow_mut(&self) -> GcCellRefMut<T>

Mutably borrows the wrapped value.

The borrow lasts until the returned GcCellRefMut exits scope. The value cannot be borrowed while this borrow is active.

Panics

Panics if the value is currently borrowed.